test: MenuGroupTest 추가
